5 rev 3 unsigned long maybe

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -11,7 +11,7 @@
 
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int i, num = 0;
+	unsigned long int i, num = 0;
 
 	i = n ^ m;
 	if (i == 0)
